+++ attachments/src/main/resources/db/migration/attachments/V5_0_1__attachment_size_bigint.sql
@@ -1,0 +1,1 @@
+alter table attachment alter column size type bigint;