+++ attachments/src/main/java/com/backstage/app/attachment/service/store/AttachmentStore.java
@@ -34,8 +34,6 @@ public interface AttachmentStore
 
 	Resource getAttachment(Attachment attachment);
 
-	Resource getAttachment(Attachment attachment, long offset, Long length);
-
 	default void saveAttachment(Attachment attachment, byte[] data)
 	{
 		saveAttachment(attachment, new ByteArrayInputStream(data));