+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 ### Dicts
 - Метод dictDataService.getByIds при передаче пустого списка идентификаторов теперь возвращает пустой список вместо исключения.
 - Добавили методы getByFilter, countByFilter, existsById, existsByFilter в генерируемый remote клиент к API.
+- Добавили принудительный каст для text[] при обработке операторов any/all.
 
 ### Jobs
 - Добавили подсветку синтаксиса cron выражений в AbstractCronJob.